refactor(#5): 컴포넌트 분리
